working on triangle only collision and correctly rendering it
maybe instead of having all the triangles in a big array, I should use a
array of objects, that contains a id and a array of tris? But that might
affect the memory layout?
Maybe i could still have all tris in a contiguous array. But save
indexes and size in a object? (would be hard and slow to insert and
remove from it, but these objects should be static.